+++ eventstore/src/options/batch_append.rs
@@ -1,7 +1,0 @@
-use eventstore_macros::{options, streaming};
-
-options! {
-    #[derive(Clone, Default)]
-    #[streaming]
-    pub struct BatchAppendOptions {}
-}