+++ kurrent/src/options/batch_append.rs
@@ -1,0 +1,7 @@
+use kurrent_macros::{options, streaming};
+
+options! {
+    #[derive(Clone, Default)]
+    #[streaming]
+    pub struct BatchAppendOptions {}
+}